requirements.txt
python-chess
requests
zstandard
pebble
